fix: rename errorFIxtures.ts to errorFixtures.ts to fix typo
